Add files via upload
The "Introduction to the Project" provides an overview of the
General Store Management System, which functions as a digital
diary for tracking expenses incurred while shopping at various
stores. Unlike traditional methods of expense tracking, this system
offers a simplified approach to monitoring spending habits. Its
primary goal is to streamline the process of managing finances
associated with general store purchases.